requirements.txt
# Webcam Capture MVP Requirements

# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0

# Audio processing
pyaudio>=0.2.11
webrtcvad>=2.0.10

# Speech recognition
openai-whisper>=20231117

# Alternative VAD (choose one)
silero-vad>=4.0.0

# Utility libraries
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
PyYAML>=6.0.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Data handling
pandas>=2.0.0
scipy>=1.11.0

# Logging and monitoring
structlog>=23.0.0
psutil>=5.9.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0